+++ src/pages/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const HomePage: React.FC = () => {
+  return <>Главная страница</>;
+};
+
+export default HomePage;
