+++ src/shared/layouts/MainLayout.tsx
@@ -3,16 +3,14 @@ import LayoutHeader from "../../widgets/LayoutHeader/LayoutHeader";
 import LayoutFooter from "../../widgets/LayoutFooter/LayoutFooter";
 import { useTheme } from "../lib/theme/useTheme";
 import "./MainLayout.css";
-interface MainLayoutProps {
-  children: React.ReactNode;
-}
+import { Outlet } from "react-router-dom";
 
-const MainLayout: React.FC<MainLayoutProps> = ({ children }) => {
+const MainLayout: React.FC = () => {
   const { theme } = useTheme();
   return (
     <main className={theme}>
       <LayoutHeader />
-      {children}
+      <Outlet />
       <LayoutFooter />
     </main>
   );
